Fixed nested filtering
Some weeks ago, and in order to be able to sort by nested values, every collection in apiserach
automatically turns a nested value. And because there was not any query using nested values, this
change broke the regular nested filtering.

Following the adopted standard, and because ALL collections are nested objects, when a new filter
is added, and if the filter field is multilevel (field.id), then a new nested filter is created
automatically, instead of a regular one.